pyproject.toml
[tool.poetry]
name = "emailsender"
version = "0.1.0"
description = ""
authors = ["IX <borntokill.ix@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
nats-py = "^2.9.0"
pydantic-settings = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
